% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply_market_model.R
\name{apply_market_model}
\alias{apply_market_model}
\title{Apply market model and return the list of objects \code{returns}}
\usage{
apply_market_model(rates, regressors, market_model = c("mean_adj", "mrkt_adj",
  "sim"), estimation_method = c("ols"), estimation_start, estimation_end)
}
\arguments{
\item{rates}{\code{list}, \code{data.frame}, \code{zoo} object containing
rates of returns of securities.}

\item{regressors}{an object of the same class as \code{rates} containing
regressors. Can be ommited, if market model is \code{mean_adj}.
\code{regressors} must have the same number of components as \code{rates}.}

\item{market_model}{a character indicating the market model among
\code{mean_adj}, \code{mrkt_adj}, and \code{sim}.}

\item{estimation_method}{a character, specifying the estimation method for
\code{sim} market model.}

\item{estimation_start}{an object of class Data, giving the start date of
estimation period.}

\item{estimation_end}{an object of class Data, giving the start date of
estimation period.}
}
\value{
A list of objects \code{returns}
}
\description{
The function applies given market model to securities' rates of returns and
returns the list of objects \code{returns} for each security, which can be
passed directly to the whole battery of tests.
}
\details{
The generic function is dispatched for such classes as \code{list},
\code{data.frame}, and \code{zoo}.
}

