% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{brown_warner_1980}
\alias{brown_warner_1980}
\title{Brown and Warner parametric test (1980).}
\usage{
brown_warner_1980(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Parametric test for event study, which is described in Brown and Warner 1980
paper.
}
\details{
Performs the parametric test for event study, which is described in Brown and
Warner 1980 paper. The test assumes cross-sectional independence and
insignificance of event-induced variance. The test examines the hypothesis
whether the theoretical cross-sectional expected value for a given day is
equal to zero. The standard deviation in statistics is calculated as the
cross-sectional mean of companies' variances, estimated on estimation period.
It calculates statistics even if event window and estimation period are
overlapped (intersect). The critical values are Student's t-distributed (no
approximation in limit). The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***). It was designed to
measure monthly data: for daily data look at Brown and Warner 1985 paper and
\code{brown_warner_1985}.
}
\examples{
## Download the historical prices for ten European insurance companies' stocks
# tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
#              "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
# prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
#                                   end = as.Date("2002-01-01"),
#                                   quote = "Close", retclass = "list")
## Estimate the rate of returns form prices
# rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
#                                compounding = "continuous")
## Download the prices and estimate the rates of market proxy (index
## ESTX50 EUR P), which is regressor for the sim model
# prices_indx <- get_prices_from_tickers("^STOXX50E",
#                                        start = as.Date("2000-01-01"),
#                                        end = as.Date("2002-01-01"),
#                                        quote = "Close", retclass = "list")
# rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
#                                     multi_day = TRUE,
#                                     compounding = "continuous")
## Apply Single Index market model
# securities_returns <- apply_market_model(rates = rates,
#                                          regressors = rates_indx,
#                                          same_regressor_for_all = TRUE,
#                                          market_model = "sim",
#                                          estimation_method = "ols",
#                                          estimation_start =
#                                                      as.Date("2001-03-26"),
#                                          estimation_end =
#                                                      as.Date("2001-09-10"))
data(securities_returns)
brown_warner_1980(securities_returns, as.Date("2001-09-11"),
                  as.Date("2001-09-28"))

}
\references{
Brown S.J., Warner J.B. \emph{Measuring security price
performance}. Journal of Financial Economics, 8:205-258, 1980.
}
\seealso{
\code{\link{parametric_tests}}, \code{\link{brown_warner_1985}},
\code{\link{t_test}}, \code{\link{patell}}, \code{\link{boehmer}}, and
\code{\link{lamb}}.
}

