% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/car_parametric_tests.R
\name{car_brown_warner_1985}
\alias{car_brown_warner_1985}
\title{Brown and Warner (1985) CAR test.}
\usage{
car_brown_warner_1985(list_of_returns, car_start, car_end, percentage)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{car_start}{the object of class \code{Date}, which represents the
first (starting) date of the time period, over which CAR will be estimated.}

\item{car_end}{the object of class \code{Date}, which represents the last
(ending) date of the time period, over which CAR will be estimated.}

\item{percentage}{the lowest allowed percentage of non-missing observation
for each day to be incorporated into CAR.}
}
\value{
The list of starting date, ending date, average percentage of
non-missing observation over each day, the value of statistics, the length
of the time period, and significance.
}
\description{
Parametric test, proposed by Brown and Warner (1995), which examines whether
or not cumulative abnormal return (CAR) significantly differs from zero.
}
\details{
This function performs the test, proposed by Brown and Warner (1985) to
investigate whether CAR significantly differs from zero. This tests uses the
variance, specified by Lamb (1995). The advantage of this test is allowance
for correlated cross-sectional returns. Test statistic is close enough to
statistic, produced by \code{\link{car_lamb}}. The critical values are
standard normal. The significance levels of \eqn{\alpha} are 0.1, 0.05, and
0.01 (marked respectively by *, **, and ***).
}
\examples{
# Download the historical prices for ten European insurance companies' stocks
tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
             "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
                                  end = as.Date("2002-01-01"),
                                  quote = "Close", retclass = "list")
# Estimate the rate of returns form prices
rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
                               compounding = "continuous")
# Download the prices and estimate the rates of market proxy (index
# ESTX50 EUR P), which is regressor for the sim model
prices_indx <- get_prices_from_tickers("^STOXX50E",
                                       start = as.Date("2000-01-01"),
                                       end = as.Date("2002-01-01"),
                                       quote = "Close", retclass = "list")
rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
                                    multi_day = TRUE,
                                    compounding = "continuous")
# Make the length of regressor and stocks to be the same
rates_indx <- replicate(n = length(tickers), expr = rates_indx)
# Apply Single Index market model
returns <- apply_market_model(rates = rates, regressors = rates_indx,
                              market_model = "sim",
                              estimation_method = "ols",
                              estimation_start = as.Date("2001-03-26"),
                              estimation_end = as.Date("2001-09-10"))
car_brown_warner_1985(list_of_returns = returns,
                      car_start = as.Date("2001-09-11"),
                      car_end = as.Date("2001-09-28"), percentage = 90)
}
\references{
Brown S.J., Warner J.B. \emph{Using Daily Stock Returns, The Case
of Event Studies}. Journal of Financial Economics, 14:3-31, 1985.
}
\seealso{
\code{\link{car_lamb}} and
\code{\link{car_parametric_tests}}.
}

