% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/car_parametric_tests.R
\name{car_parametric_tests}
\alias{car_parametric_tests}
\title{Returns the result of given parametric CAR tests for event study.}
\usage{
car_parametric_tests(list_of_returns, car_start, car_end, percentage,
  all = TRUE, tests)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{car_start}{the object of class \code{Date}, which represents the
first (starting) date of the time period, over which CAR will be estimated.}

\item{car_end}{the object of class \code{Date}, which represents the last
(ending) date of the time period, over which CAR will be estimated.}

\item{percentage}{the lowest allowed percentage of non-missing observation
for each day to be incorporated into CAR.}

\item{all}{a logical value indicating whether all tests should be performed.
The default value is \code{TRUE}.}

\item{tests}{the list of tests functions among \code{car_brown_warner_1985}
and \code{car_lamb}.}
}
\value{
The list of results of given tests.
}
\description{
Performs given tests to examine whether cumulative abnormal return (CAR)
significantly differs from zero.
}
\details{
\code{car_parametric_tests} performs specified tests among
\code{car_brown_warner_1985} and \code{lamb} and returns the list of these
tests' results. If \code{all = TRUE} (by default), the function ignores the
value of \code{tests}.
}
\references{
\itemize{
\item Brown S.J., Warner J.B. \emph{Using Daily Stock Returns, The Case of
Event Studies}. Journal of Financial Economics, 14:3-31, 1985.
\item Lamb R.P. \emph{An Exposure-Based Analysis of Property-Liability
Insurer Stock Values around Hurricane Andrew}. Journal of Risk and Insurance,
62(1):111-123, 1995.}
}
\seealso{
\code{\link{car_brown_warner_1985}} and \code{\link{car_lamb}}.
}

