% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_prices_from_tickers.R
\name{get_prices_form_tickers}
\alias{get_prices_form_tickers}
\title{Daily Prices of Securities}
\usage{
get_prices_form_tickers(..., start, end, quote = c("Open", "Close"),
  retclass = c("list", "zoo", "date.frame"))
}
\arguments{
\item{...}{character vectors indicating the tickers (should be valid in Yahoo
Finance).}

\item{start}{the object of the Date class, which specifies the first date of
observed time period.}

\item{end}{the object of the Date class, which specifies the last date of
observed time period.}

\item{quote}{a character indicating the type of the price (Open or Close).
The default value is Open}

\item{retclass}{a character specifying the return class. "list", "zoo"
, or "data.frame", by default is "list"}
}
\value{
the data about prices in given class container: "list", "zoo",
 or "data.frame".
}
\description{
Returns the daily Open or Close prices between \code{start} and \code{end}
date for given tickers.
}
\details{
This function uses the function \code{get.hist.quote} form packege
\code{tseries}. The providor set automatically to Yahoo Finance. The function
returns the data in different class-conrainers: list of zoo's, zoo, or
data.frame.
}
\seealso{
\code{\link[tseries]{get.hist.quote}}
}

