% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prices_from_tickers.R
\name{get_prices_from_tickers}
\alias{get_prices_from_tickers}
\title{Returns daily prices for securities.}
\usage{
get_prices_from_tickers(..., start, end, quote = c("Open", "Close"),
  retclass = c("list", "zoo", "data.frame"))
}
\arguments{
\item{...}{character vectors indicating tickers (should be valid in Yahoo
Finance).}

\item{start}{an object of Date class, which specifies the first date of the
observed time period.}

\item{end}{an object of Date class, which specifies the last date of the
observed time period.}

\item{quote}{a character indicating the type of the price (Open or Close).
The default value is set to Open.}

\item{retclass}{a character specifying the return class: "list", "zoo" or
"data.frame". The default value is set to "list".}
}
\value{
Prices of securities as "list", "zoo" or "data.frame".
}
\description{
Returns daily Open or Close prices between \code{start} and \code{end}
date for given tickers.
}
\details{
This function uses the function \code{get.hist.quote} form the package
\code{tseries}. The provider is set automatically to Yahoo Finance. The
 function returns the data in different class-containers: list of zoo's, zoo,
 or data.frame.
}
\examples{
## Download the historical prices for ten European insurance companies'
stocks:
\dontrun{
tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
             "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
                                  end = as.Date("2002-01-01"),
                                  quote = "Close", retclass = "list")
}
## The result of the above code is stored in:
data(prices)

}
\seealso{
\code{\link[tseries]{get.hist.quote}}
}
