% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{lamb}
\alias{lamb}
\title{Lamb's parametric test (1995).}
\usage{
lamb(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Parametric test for event study, which is descibed in Lamb's 1995
paper.
}
\details{
Performs the parametric test for event study, which is descibed in Lamb's
1995 paper. The author refers to Warner and Brown's (1985) and Henderson's
(1990) paper, however this test was not observed in niether papers. The test
statistics are very close to the statistics produced by
\code{brown_warner_1985} and typically has the same significance. The test
examines the hypothesis whether the theoretical cross-sectional expected
value for a given day is equal to zero. It calculates statistics even if
event window and estimation period are overlapped (intersect). The critical
values are standard normal. The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***).
}
\references{
Lamb R.P. \emph{An Exposure-Based Analysis of Property-Liability
Insurer Stock Values around Hurricane Andrew}. Journal of Risk and Insurance,
62(1):111-123, 1995.
}
\seealso{
\code{\link{parametric_tests}}, \code{\link{brown_warner_1980}},
\code{\link{brown_warner_1985}}, \code{\link{t_test}},\code{\link{patell}}
and \code{\link{boehmer}}
}

