% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{nonparametric_tests}
\alias{nonparametric_tests}
\title{Returns the result for all nonparametric tests for event study.}
\usage{
nonparametric_tests(list_of_returns, event_start, event_end, all = TRUE,
  tests)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each
element of which is treated as a company (security).}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}

\item{all}{a logical value indicating whether all tests should be performed.
The default value is \code{TRUE}.}

\item{tests}{the list of tests functions among \code{\link{sign_test}},
\code{\link{generalized_sign_test}}, \code{\link{corrado_sign_test}},
\code{\link{rank_test}}, \code{\link{modified_rank_test}}, and
\code{\link{wilcoxon_test}}.}
}
\value{
The single table of statistics and significances of all tests.
}
\description{
Performs main nonparametric tests for each date in the event window and
returns the table of statistics and significance of these tests.
}
\details{
\code{nonparametric_tests} performs given tests among \code{\link{sign_test}},
\code{\link{generalized_sign_test}}, \code{\link{corrado_sign_test}},
\code{\link{rank_test}}, \code{\link{modified_rank_test}},
\code{\link{wilcoxon_test}}, and merge result to single table. If
\code{all = TRUE} (by default), the function ignores the value of \code{tests}.
}
\examples{
# Download the historical prices for ten European insurance companies' stocks
tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
             "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
                                  end = as.Date("2002-01-01"),
                                  quote = "Close", retclass = "list")
# Estimate the rate of returns form prices
rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
                               compounding = "continuous")
# Download the prices and estimate the rates of market proxy (index
# ESTX50 EUR P), which is regressor for the sim model
prices_indx <- get_prices_from_tickers("^STOXX50E",
                                       start = as.Date("2000-01-01"),
                                       end = as.Date("2002-01-01"),
                                       quote = "Close", retclass = "list")
rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
                                    multi_day = TRUE,
                                    compounding = "continuous")
# Make the length of regressor and stocks to be the same
rates_indx <- replicate(n = length(tickers), expr = rates_indx)
# Apply Single Index market model
returns <- apply_market_model(rates = rates, regressors = rates_indx,
                              market_model = "sim",
                              estimation_method = "ols",
                              estimation_start = as.Date("2001-03-26"),
                              estimation_end = as.Date("2001-09-10"))
nonparametric_tests(returns, as.Date("2001-09-11"), as.Date("2001-09-28"))
}
\references{
\itemize{
\item Corrado C.J., Zivney T.L. \emph{The Specification and Power of
the Sign Test in Event Study Hypothesis Tests Using Daily Stock Returns}.
Journal of Financial and Quantitative Analysis, 27(3):465-478, 1992.
\item McConnell J.J., Muscarella C.J. \emph{Capital expenditure plans and
firm value} Journal of Financial Economics, 14:399-422, 1985.
\item Boehmer E., Musumeci J., Poulsen A.B. \emph{Event-study methodology
under conditions of event-induced variance}. Journal of Financial Economics,
30(2):253-272, 1991.
\item Cowan A.R. \emph{Nonparametric Event Study Tests}. Review of
Quantitative Finance and Accounting, 2:343-358, 1992.
\item Corrado C.J. \emph{A Nonparametric Test for Abnormal Security-Price
Performance in Event Studies}. Journal of Financial Economics 23:385-395,
1989.
\item Campbell C.J., Wasley C.E. \emph{Measuring Security Price Performance
Using Daily NASDAQ Returns}. Journal of Financial Economics 33:73-92, 1993.
\item Savickas R. \emph{Event-Induced Volatility and Tests for Abnormal
Performance}. The Journal of Financial Research, 26(2):156-178, 2003.
\item Kolari J.W., Pynnonen S. \emph{Event Study Testing with Cross-sectional
Correlation of Abnormal Returns}. The Review of Financial Studies,
23(11):3996-4025, 2010.
\item Wilcoxon F. \emph{Individual Comparisons by Ranking Methods}.
Biometrics Bulletin 1(6):80-83, 1945.
\item Lehmann E.L, \emph{Nonparametrics: Statistical Methods Based on Ranks}.
San Francisco: Holden-Day, 1975.
\item Hollander M., Wolfe D.A. \emph{Nonparametric Statistical Methods}.
New York: John Wiley & Sons, 1973.
}
}
\seealso{
\code{\link{sign_test}}, \code{\link{generalized_sign_test}},
\code{\link{corrado_sign_test}}, \code{\link{rank_test}},
\code{\link{modified_rank_test}}, and \code{\link{wilcoxon_test}}.
}

