% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{parametric_tests}
\alias{parametric_tests}
\title{Returns the result of given parametric tests for event study.}
\usage{
parametric_tests(list_of_returns, event_start, event_end, all = TRUE, tests)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}

\item{all}{a logical value indicating whether all tests should be performed.
The default value is \code{TRUE}.}

\item{tests}{the list of tests functions among \code{brown_warner_1980},
\code{brown_warner_1985}, \code{t_test}, \code{patell}, \code{boehmer}, and
\code{lamb}.}
}
\value{
The single table of statistics and significances of all tests.
}
\description{
Performs main parametric tests for each date in the event window and returns
the table of statistics and significance.
}
\details{
\code{parametric_tests} performs given tests among \code{brown_warner_1980},
\code{brown_warner_1985}, \code{t_test}, \code{patell}, \code{boehmer},
\code{lamb} and merge result to single table. If \code{all = TRUE} (by default),
the function ignores the value of \code{tests}.
}
\examples{
## Download the historical prices for ten European insurance companies' stocks
# tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
#              "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
# prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
#                                   end = as.Date("2002-01-01"),
#                                   quote = "Close", retclass = "list")
## Estimate the rate of returns form prices
# rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
#                                compounding = "continuous")
## Download the prices and estimate the rates of market proxy (index
## ESTX50 EUR P), which is regressor for the sim model
# prices_indx <- get_prices_from_tickers("^STOXX50E",
#                                        start = as.Date("2000-01-01"),
#                                        end = as.Date("2002-01-01"),
#                                        quote = "Close", retclass = "list")
# rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
#                                     multi_day = TRUE,
#                                     compounding = "continuous")
## Apply Single Index market model
# securities_returns <- apply_market_model(rates = rates,
#                                          regressors = rates_indx,
#                                          same_regressor_for_all = TRUE,
#                                          market_model = "sim",
#                                          estimation_method = "ols",
#                                          estimation_start =
#                                                      as.Date("2001-03-26"),
#                                          estimation_end =
#                                                      as.Date("2001-09-10"))
data(securities_returns)
parametric_tests(securities_returns, as.Date("2001-09-11"),
                 as.Date("2001-09-28"))
}
\references{
\itemize{
\item Brown S.J., Warner J.B. \emph{Measuring security price performance}.
Journal of Financial Economics, 8:205-258, 1980.
\item Brown S.J., Warner J.B. \emph{Using Daily Stock Returns, The Case of
Event Studies}. Journal of Financial Economics, 14:3-31, 1985.
\item Boehmer, E. \emph{Event-study methodology under conditions of event-
induced variance}. Journal of Financial Economics, 30(2):253-272, 1991.
\item Patell J.M. \emph{Corporate forecasts of earnings per share and stock
price behavior: empirical tests}. Journal of Accounting Research, 14(2):246-
276, 1976.
\item Lamb R.P. \emph{An Exposure-Based Analysis of Property-Liability
Insurer Stock Values around Hurricane Andrew}. Journal of Risk and Insurance,
62(1):111-123, 1995.}
}
\seealso{
\code{\link{brown_warner_1980}}, \code{\link{brown_warner_1985}},
\code{\link{t_test}}, \code{\link{patell}}, \code{\link{boehmer}}, and
\code{\link{lamb}}.
}

