% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices_indx}
\alias{prices_indx}
\title{Prices of S&P 500 index from 2019-04-01 to 2020-04-01}
\format{
A \code{zoo} object.
}
\usage{
prices_indx
}
\description{
A \code{zoo} object containing daily prices of S&P 500 index from
2019-04-01 to 2020-04-01. See examples of
\code{\link{get_prices_from_tickers}} for the dataset generation.
}
\keyword{datasets}
