% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{Rates of returns of nine major insurance companies from 2000-01-04 to
2002-01-01}
\format{A \code{zoo} object of nine columns:
\itemize{
  \item ALV.DE
  \item CS.PA
  \item ELE.PA
  \item G.MI
  \item HNR1.HA
  \item HSX.L
  \item MUV2.DE
  \item RSA.L
  \item TOP.CO
}}
\usage{
rates
}
\description{
A \code{zoo} object of nine columns containing  daily rates of returns of
nine major insurance companies from 2000-01-04 to 2002-01-01. The code for
generatin this dataset can be seen in the documentation of
\code{get_rates_from_prices} function.
}
\keyword{datasets}
