% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny demo app}
\usage{
run_app()
}
\description{
The function \code{run_app()} launches a Shiny app, which is a GUI wrapper
of crippled version of \code{{estudy2}}. This app is developed exclusively
for demonstration purposes.
}
\details{
The app is run locally.
}
