% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo.R
\name{run_demo}
\alias{run_demo}
\title{Run Shiny demo app}
\usage{
run_demo()
}
\description{
The function \code{run_demo()} launches a Shiny app, which is a GUI wrapper
of crippled version of \code{{estudy2}}. This app is developed exclusively
for demonstration purposes.
}
\details{
The app is run locally.
}
