% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{wilcoxon_test}
\alias{wilcoxon_test}
\title{Wilcoxon signed rank test for event study.}
\usage{
wilcoxon_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Performs Wilcoxon test on event period for abnormal returns (the latter are
treated as differences).
}
\details{
The estimation periods can overlap with event windows, because the procedure
takes into account only abnormal returns from event window. The test has the
same algorithm as built-in \code{R} \code{\link{wilcox.test}}. The critical
values are exact values, which are obtained from \code{\link{qsignrank}}. The
algorithm is the following: for each day in event window the cross-sectional
abnormal returns treated as sample of differences. Firstly the absolute value
of these differences are computed, and corresponding ranks of non-zero values
are calculated. The test statistic is the sum of ranks, corresponding to
positive abnormal returns. The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***).
}
\examples{
## Download the historical prices for ten European insurance companies' stocks
# tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
#              "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
# prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
#                                   end = as.Date("2002-01-01"),
#                                   quote = "Close", retclass = "list")
## Estimate the rate of returns form prices
# rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
#                                compounding = "continuous")
## Download the prices and estimate the rates of market proxy (index
## ESTX50 EUR P), which is regressor for the sim model
# prices_indx <- get_prices_from_tickers("^STOXX50E",
#                                        start = as.Date("2000-01-01"),
#                                        end = as.Date("2002-01-01"),
#                                        quote = "Close", retclass = "list")
# rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
#                                     multi_day = TRUE,
#                                     compounding = "continuous")
## Apply Single Index market model
# securities_returns <- apply_market_model(rates = rates,
#                                          regressors = rates_indx,
#                                          same_regressor_for_all = TRUE,
#                                          market_model = "sim",
#                                          estimation_method = "ols",
#                                          estimation_start =
#                                                      as.Date("2001-03-26"),
#                                          estimation_end =
#                                                      as.Date("2001-09-10"))
data(securities_returns)
wilcoxon_test(securities_returns, as.Date("2001-09-11"),
              as.Date("2001-09-28"))

}
\references{
\itemize{
\item Wilcoxon F. \emph{Individual Comparisons by Ranking Methods}.
Biometrics Bulletin 1(6):80-83, 1945.
\item Kolari J.W., Pynnonen S. \emph{Event Study Testing with Cross-sectional
Correlation of Abnormal Returns}. The Review of Financial Studies,
23(11):3996-4025, 2010.
\item Lehmann E.L, \emph{Nonparametrics: Statistical Methods Based on Ranks}.
San Francisco: Holden-Day, 1975.
\item Hollander M., Wolfe D.A. \emph{Nonparametric Statistical Methods}.
New York: John Wiley & Sons, 1973.
}
}
\seealso{
\code{\link{nonparametric_tests}}, \code{\link{sign_test}},
\code{\link{generalized_sign_test}}, \code{\link{corrado_sign_test}},
\code{\link{rank_test}}, and \code{\link{modified_rank_test}}.
}
