% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply_market_model.R
\name{apply_market_model}
\alias{apply_market_model}
\title{Apply market model and return the list of objects \code{returns}}
\usage{
apply_market_model(rates, regressors, market_model = c("mean_adj", "mrkt_adj",
  "sim"), estimation_method = c("ols"), estimation_start, estimation_end)
}
\arguments{
\item{rates}{\code{list}, \code{data.frame}, \code{zoo} object containing
rates of returns of securities.}

\item{regressors}{an object of the same class as \code{rates} containing
regressors. Can be ommited, if market model is \code{mean_adj}.
\code{regressors} must have the same number of components as \code{rates}.}

\item{market_model}{a character indicating the market model among
\code{mean_adj}, \code{mrkt_adj}, and \code{sim}.}

\item{estimation_method}{a character, specifying the estimation method for
\code{sim} market model.}

\item{estimation_start}{an object of class Data, giving the start date of
estimation period.}

\item{estimation_end}{an object of class Data, giving the start date of
estimation period.}
}
\value{
A list of objects \code{returns}
}
\description{
The function applies given market model to securities' rates of returns and
returns the list of objects \code{returns} for each security, which can be
passed directly to the whole battery of tests.
}
\details{
The generic function is dispatched for such classes as \code{list},
\code{data.frame}, and \code{zoo}.
}
\examples{
# Download the historical prices for ten European insurance companies' stocks
tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
             "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
                                  end = as.Date("2002-01-01"),
                                  quote = "Close", retclass = "list")
# Estimate the rate of returns form prices
rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
                               compounding = "continuous")
### Mean-adjusted returns market model
# Apply mean-adjusted returns market model for each company
returns <- apply_market_model(rates, market_model = "mean_adj",
                              estimation_start = as.Date("2001-03-26"),
                              estimation_end = as.Date("2001-09-10"))
### Single Index market model
# Download the prices and estimate the rates of market proxy (index
# ESTX50 EUR P), which is regressor for the sim model
prices_indx <- get_prices_from_tickers("^STOXX50E",
                                       start = as.Date("2000-01-01"),
                                       end = as.Date("2002-01-01"),
                                       quote = "Close", retclass = "list")
rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
                                    multi_day = TRUE,
                                    compounding = "continuous")
# Make the length of regressor and stocks to be the same
rates_indx <- replicate(n = length(tickers), expr = rates_indx)
# Apply Single Index market model
returns <- apply_market_model (rates = rates, regressors = rates_indx,
                               market_model = "sim",
                               estimation_method = "ols",
                               estimation_start = as.Date("2001-03-26"),
                               estimation_end = as.Date("2001-09-10"))
}

