% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{boehmer}
\alias{boehmer}
\title{Boehmer's parametric test (1991).}
\usage{
boehmer(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each elemnt
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Parametric test for event study, which is descibed in Boehmer's 1991
paper.
}
\details{
Performs the parametric test for event study, which is descibed in Boehmer's
1991 paper, also called hybrid test or standardized cross-sectional test.
This test performs t-test based on Patell's standardized resuduals. By
combining Patell's and t- tests this test allows event-induced variance
changes, but still assumes cross-sectional independence. The test examines
the hypothesis whether the theoretical cross-sectional expected value for a
given day is equal to zero. It calculates statistics even if event window and
estimation period are overlapped (intersect). The critical values has
Student's t-distribution. The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***).
}
\references{
\itemize{
\item Patell J.M. \emph{Corporate forecasts of earnings per share and stock
price behavior: empirical tests}. Journal of Accounting Research, 14(2):246-
276, 1976.
\item Boehmer, E. \emph{
Event-study methodology under conditions of event-
induced variance}. Journal of Financial Economics, 30(2):253-272, 1991.}
}
\seealso{
\code{\link{parametric_tests}}, \code{\link{brown_warner_1980}},
\code{\link{brown_warner_1985}}, \code{\link{t_test}}, and
\code{\link{patell}}
}

