% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{brown_warner_1980}
\alias{brown_warner_1980}
\title{Brown and Warner parametric test (1980).}
\usage{
brown_warner_1980(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Parametric test for event study, which is descibed in Brown and Warner 1980
paper.
}
\details{
Performs the parametric test for event study, which is descibed in Brown and
Warner 1980 paper. The test assumes cross-sectional independence and
insignificance of event-induced variance. The test examines the hypothesis
whether the theoretical cross-sectional expected value for a given day is
equal to zero. The standard deviation in statistics is calculated as the
cross-sectional mean of companies' variances, estimated on estimation period.
It calculates statistics even if event window and estimation period are
overlapped (intersect). The critical values are Student's t-destributed (no
approximation in limit). The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***). It was designed to
measure monthly data: for daily data look at Brown and Warner 1985 paper and
\code{brown_warner_1985}.
}
\references{
Brown S.J., Warner J.B. \emph{Measuring security price
performance}. Journal of Financial Economics, 8:205-258, 1980.
}
\seealso{
\code{\link{parametric_tests}}, \code{\link{brown_warner_1985}},
\code{\link{t_test}}, \code{\link{patell}}, and \code{\link{boehmer}}
}

