% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/car_parametric_tests.R
\name{car_brown_warner_1985}
\alias{car_brown_warner_1985}
\title{Brown and Warner (1985) CAR test.}
\usage{
car_brown_warner_1985(list_of_returns, car_start, car_end, percentage)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{car_start}{the object of class \code{Date}, which represents the
first (starting) date of the time period, over which CAR will be estimated.}

\item{car_end}{the object of class \code{Date}, which represents the last
(ending) date of the time period, over which CAR will be estimated.}

\item{percentage}{the lowest allowed percentage of non-missing observation
for each day to be incorporated into CAR.}
}
\value{
The list of starting date, ending date, average percentage of
non-missing observation over each day, the value of statistics, the length
of the time period, and significance.
}
\description{
Parametric test, proposed by Brown and Warner (1995), which examines whether
or not cumulative abnormal return (CAR) significantly differs from zero.
}
\details{
This function performs the test, proposed by Brown and Warner (1985) to
investigate whether CAR significantly differs from zero. This tests uses the
variance, specified by Lamb (1995). The advantage of this test is allowance
for correlated cross-sectional returns. Test statistic is close enough to
statistic, produced by \code{\link{car_lamb}}. The critical values are
standard normal. The significance levels of \eqn{\alpha} are 0.1, 0.05, and
0.01 (marked respectively by *, **, and ***).
}
\references{
Brown S.J., Warner J.B. \emph{Using Daily Stock Returns, The Case
of Event Studies}. Journal of Financial Economics, 14:3-31, 1985.
}
\seealso{
\code{\link{car_lamb}} and
\code{\link{car_parametric_tests}}.
}

