% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{corrado_sign_test}
\alias{corrado_sign_test}
\title{Corrado's sign test (1992).}
\usage{
corrado_sign_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
The implementation of nonparametric test, described in Corrado and Zivney's
1992 paper.
}
\details{
Performs the nonparametric test for event study, which is descibed in Corrado
and Zivney's 1992 paper. This test is similar to procedure, described in
Brown and Warner's paper 1985 (t-ratio), but instead of using abnormal
returns, the test uses \eqn{G_{i,t} = sign(A_{i,t} - median(A_i))}.
\code{sign} and \code{median} are ones, which have the same definition as R
functions. For this test the estimation period and event period must not
overlap, otherwise an error occurrs. The sign test procedure avoids the
misspecification of tests, which assume symmetry around zero of abnormal
returns (the median equals to zero). For a single day the performance of this
test is proven to be better than classical Brown and Warner's test (without
event-induced variance). This test is dominated by rank test. The
significance levels of \eqn{\alpha} are 0.1, 0.05, and 0.01 (marked
respectively by *, **, and ***).
}
\references{
Corrado C.J., Zivney T.L. \emph{The Specification and Power of
the Sign Test in Event Study Hypothesis Tests Using Daily Stock Returns}.
Journal of Financial and Quantitative Analysis, 27(3):465-478, 1992.
}
\seealso{
\code{\link{nonparametric_tests}}, \code{\link{sign_test}},
\code{\link{generalized_sign_test}}, \code{\link{rank_test}},
\code{\link{modified_rank_test}}, and \code{\link{wilcoxon_test}}.
}

