% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{generalized_sign_test}
\alias{generalized_sign_test}
\title{Binomial sign test for event study.}
\usage{
generalized_sign_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each elemnt
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
The binomial sign test which determines whether the frequency of positive
abnormal returns in the event period is significantly different from one in
the estimation period.
}
\details{
This test is application of the simple binomial test to the event study,
which indicates whether the cross-sectional frequency of positive abnormal
returns is significantly different from the the expected. This test is stable
to outliers, in other words allows to check if the result is driven by few
companies with extremly large abnomal performance. For this test the
estimation period and event period must not overlap, otherwise an error
occurrs. This test instead of using naive value of expected frequency 0.5
uses an estimate from the estimation period. The test statistic is assumed to
have a normal distribution. Typically is used together with parametric tests.
The test is well-specified for the case, when cross-sectional abnoral returns
are not symmetric. Also this procedure is less sensative to extreme returns
than the rank test. The significance levels of \eqn{\alpha} are 0.1, 0.05,
and 0.01 (marked respectively by *, **, and ***).
}
\references{
\itemize{
\item McConnell J.J., Muscarella C.J. \emph{Capital expenditure plans and
firm value} Journal of Financial Economics, 14:399-422, 1985.
\item Boehmer E., Musumeci J., Poulsen A.B. \emph{Event-study methodology
under conditions of event-induced variance}. Journal of Financial Economics,
30(2):253-272, 1991.
\item Cowan A.R. \emph{Nonparametric Event Study Tests}. Review of
Quantitative Finance and Accounting, 2:343-358, 1992.
}
}
\seealso{
\code{\link{nonparametric_tests}}, \code{\link{corrado_sign_test}},
\code{\link{rank_test}}, and \code{\link{modified_rank_test}}
}

