% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prices_from_tickers.R
\name{get_prices_from_tickers}
\alias{get_prices_from_tickers}
\title{Returns the daily prices for securities.}
\usage{
get_prices_from_tickers(..., start, end, quote = c("Open", "Close"),
  retclass = c("list", "zoo", "data.frame"))
}
\arguments{
\item{...}{character vectors indicating the tickers (should be valid in Yahoo
Finance).}

\item{start}{the object of the Date class, which specifies the first date of
observed time period.}

\item{end}{the object of the Date class, which specifies the last date of
observed time period.}

\item{quote}{a character indicating the type of the price (Open or Close).
The default value is Open}

\item{retclass}{a character specifying the return class. "list", "zoo"
, or "data.frame", by default is "list"}
}
\value{
the data about prices in given class container: "list", "zoo",
 or "data.frame".
}
\description{
Returns the daily Open or Close prices between \code{start} and \code{end}
date for given tickers.
}
\details{
This function uses the function \code{get.hist.quote} form package
\code{tseries}. The provider set automatically to Yahoo Finance. The function
returns the data in different class-containers: list of zoo's, zoo, or
data.frame.
}
\examples{
## Download the historical prices for ten European insurance companies' stocks
# tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
#              "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
# prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
#                                   end = as.Date("2002-01-01"),
#                                   quote = "Close", retclass = "list")
data(prices)

}
\seealso{
\code{\link[tseries]{get.hist.quote}}
}

