% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{lamb}
\alias{lamb}
\title{Lamb's parametric test (1995).}
\usage{
lamb(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Parametric test for event study, which is descibed in Lamb's 1995
paper.
}
\details{
Performs the parametric test for event study, which is descibed in Lamb's
1995 paper. The author refers to Warner and Brown's (1985) and Henderson's
(1990) paper, however this test was not observed in niether papers. The test
statistics are very close to the statistics produced by
\code{brown_warner_1985} and typically has the same significance. The test
examines the hypothesis whether the theoretical cross-sectional expected
value for a given day is equal to zero. It calculates statistics even if
event window and estimation period are overlapped (intersect). The critical
values are standard normal. The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***).
}
\examples{
# Download the historical prices for ten European insurance companies' stocks
tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
             "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
                                  end = as.Date("2002-01-01"),
                                  quote = "Close", retclass = "list")
# Estimate the rate of returns form prices
rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
                               compounding = "continuous")
# Download the prices and estimate the rates of market proxy (index
# ESTX50 EUR P), which is regressor for the sim model
prices_indx <- get_prices_from_tickers("^STOXX50E",
                                       start = as.Date("2000-01-01"),
                                       end = as.Date("2002-01-01"),
                                       quote = "Close", retclass = "list")
rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
                                    multi_day = TRUE,
                                    compounding = "continuous")
# Make the length of regressor and stocks to be the same
rates_indx <- replicate(n = length(tickers), expr = rates_indx)
# Apply Single Index market model
returns <- apply_market_model(rates = rates, regressors = rates_indx,
                              market_model = "sim",
                              estimation_method = "ols",
                              estimation_start = as.Date("2001-03-26"),
                              estimation_end = as.Date("2001-09-10"))
lamb(returns, as.Date("2001-09-11"), as.Date("2001-09-28"))
}
\references{
Lamb R.P. \emph{An Exposure-Based Analysis of Property-Liability
Insurer Stock Values around Hurricane Andrew}. Journal of Risk and Insurance,
62(1):111-123, 1995.
}
\seealso{
\code{\link{parametric_tests}}, \code{\link{brown_warner_1980}},
\code{\link{brown_warner_1985}}, \code{\link{t_test}},\code{\link{patell}}
and \code{\link{boehmer}}.
}

