% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{modified_rank_test}
\alias{modified_rank_test}
\title{Modified rank test for event study.}
\usage{
modified_rank_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each elemnt
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
This test, the modification of the original rank test, proposed by Corrado
(1989), is abapted to missing values in abnormal returns.
}
\details{
In addition to original rank test, the procedure devides corresponding ranks
on number of nonmissing returns plus one for each security. This leads to
order statistics with uniform destribution. In limit overall statistics under
null hypothesis is aproximately noramly distributed. For this test the
estimation period and event period must not overlap, otherwise an error
occurrs. The test is well-specified for the case, when cross-sectional
abnoral returns are not symmetric. The test is stable to variance increase
during event window. This test is more sensative to extreme values than sign
test. The significance levels of \eqn{\alpha} are 0.1, 0.05, and 0.01 (marked
respectively by *, **, and ***).
}
\references{
\itemize{
\item Corrado C.J., Zivney T.L. \emph{The Specification and Power of
the Sign Test in Event Study Hypothesis Tests Using Daily Stock Returns}.
Journal of Financial and Quantitative Analysis, 27(3):465-478, 1992.
\item Kolari J.W., Pynnonen S. \emph{Event Study Testing with Cross-sectional
Correlation of Abnormal Returns}. The Review of Financial Studies,
23(11):3996-4025, 2010.
}
}
\seealso{
\code{\link{nonparametric_tests}}, \code{\link{corrado_sign_test}},
\code{\link{binomial_sign_test}}, and \code{\link{rank_test}}
}

