% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices}
\alias{prices}
\title{Stock prices of seven companies from 2019-04-01 to 2020-04-01}
\format{
A \code{zoo} object of eight columns:
\itemize{
  \item AMZN
  \item ZM
  \item UBER
  \item NFLX
  \item SHOP
  \item FB
  \item UPWK
}
}
\usage{
prices
}
\description{
A \code{zoo} object of seven columns containing daily Close stock prices from
2019-04-01 to 2020-04-01 of seven companies, which could profit from COVID-19
lockdown. See examples of \code{\link{get_prices_from_tickers}} for the
dataset generation.
}
\keyword{datasets}
