% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices_indx}
\alias{prices_indx}
\title{Prices of of ESTX50 EUR P index from 2000-01-03 to 2001-12-31}
\format{A \code{zoo} object.}
\usage{
prices_indx
}
\description{
A \code{zoo} object containing daily prices of ESTX50 EUR P index from
2000-01-03 to 2001-12-31. The code for generatin this dataset can be seen in
the documentation of \code{get_prices_from_tickers} function.
}
\keyword{datasets}
