% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{rank_test}
\alias{rank_test}
\title{Rank test for event study.}
\usage{
rank_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
The original rank test for event study, which is based on Wilcoxon (1945)
rank test.
}
\details{
This procedure uses ranks of abnormal returns to examine significance of each
day in event window. In order to get ranks of corresponding abnormal returns,
the procedure uses regular R function \code{\link{rank}} with parameter
\code{ties.method = "average"} and \code{na.last = "keep"}. For this test the
estimation period and event period must not overlap, otherwise an error
occurs. The test statistic is assumed to have a normal distribution (as an
approximation). The test is well-specified for the case, when cross-sectional
abnormal returns are not symmetric. The test is stable to variance increase
during event window. This test is more sensitive to extreme values than sign
test. For data with missed data see the \code{\link{modified_rank_test}}. The
significance levels of \eqn{\alpha} are 0.1, 0.05, and 0.01 (marked
respectively by *, **, and ***).
}
\examples{
## Download the historical prices for ten European insurance companies' stocks
# tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
#              "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
# prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
#                                   end = as.Date("2002-01-01"),
#                                   quote = "Close", retclass = "list")
## Estimate the rate of returns form prices
# rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
#                                compounding = "continuous")
## Download the prices and estimate the rates of market proxy (index
## ESTX50 EUR P), which is regressor for the sim model
# prices_indx <- get_prices_from_tickers("^STOXX50E",
#                                        start = as.Date("2000-01-01"),
#                                        end = as.Date("2002-01-01"),
#                                        quote = "Close", retclass = "list")
# rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
#                                     multi_day = TRUE,
#                                     compounding = "continuous")
## Apply Single Index market model
# securities_returns <- apply_market_model(rates = rates,
#                                          regressors = rates_indx,
#                                          same_regressor_for_all = TRUE,
#                                          market_model = "sim",
#                                          estimation_method = "ols",
#                                          estimation_start =
#                                                      as.Date("2001-03-26"),
#                                          estimation_end =
#                                                      as.Date("2001-09-10"))
data(securities_returns)
rank_test(securities_returns, as.Date("2001-09-11"), as.Date("2001-09-28"))
}
\references{
\itemize{
\item Corrado C.J. \emph{A Nonparametric Test for Abnormal Security-Price
Performance in Event Studies}. Journal of Financial Economics 23:385-395,
1989.
\item Cowan A.R. \emph{Nonparametric Event Study Tests}. Review of
Quantitative Finance and Accounting, 2:343-358, 1992.
\item Campbell C.J., Wasley C.E. \emph{Measuring Security Price Performance
Using Daily NASDAQ Returns}. Journal of Financial Economics 33:73-92, 1993.
\item Savickas R. \emph{Event-Induced Volatility and Tests for Abnormal
Performance}. The Journal of Financial Research, 26(2):156-178, 2003.
}
}
\seealso{
\code{\link{nonparametric_tests}},\code{\link{sign_test}},
\code{\link{generalized_sign_test}}, \code{\link{corrado_sign_test}},
\code{\link{modified_rank_test}}, and \code{\link{wilcoxon_test}}.
}

