% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{Rates of returns of 9 major insurance companies from 2000-01-01 to 2002-01-01}
\format{A list of 10 \code{zoo} elements:
\itemize{
  \item ALV.DE
  \item CS.PA
  \item ELE.PA
  \item G.MI
  \item HNR1.HA
  \item HSX.L
  \item MUV2.DE
  \item RSA.L
  \item TOP.CO
}}
\usage{
rates
}
\description{
A list, which contains 9 elements, each of which is \code{zoo} object,
containing daily rates of returns of corresponding 9 major insurance
companies from 2000-01-01 to 2002-01-01. The code for generatin this dataset
can be seen in the documentation of \code{get_rates_from_prices} function.
The variables are as follows:
}
\keyword{datasets}
