% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates}
\alias{rates}
\title{Rates of returns of seven companies from 2019-04-01 to 2020-04-01}
\format{
A \code{zoo} object of eight columns:
\itemize{
  \item AMZN
  \item ZM
  \item UBER
  \item NFLX
  \item SHOP
  \item FB
  \item UPWK
}
}
\usage{
rates
}
\description{
A \code{zoo} object of seven columns containing daily rates of returns from
2019-04-01 to 2020-04-01 of seven companies, which could profit from COVID-19
lockdown. See examples of \code{\link{get_rates_from_prices}} for the dataset
generation.
}
\keyword{datasets}
