% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates_indx}
\alias{rates_indx}
\title{Rates of returns of ESTX50 EUR P index from 2000-01-01 to 2002-01-01}
\format{A list of 1 zoo object.}
\usage{
rates_indx
}
\description{
A list of one element, which is presented by \code{zoo} class and contains
daily rates of returns of ESTX50 EUR P index from 2000-01-01 to 2002-01-01.
The code for generatin this dataset can be seen in the example of
\code{apply_market_model} function.
}
\keyword{datasets}

