% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rates_indx}
\alias{rates_indx}
\title{Rates of returns of EURONEXT 100 index from 2000-01-04 to 2001-12-31}
\format{
A \code{zoo} object.
}
\usage{
rates_indx
}
\description{
A \code{zoo} object containing daily rates of returns of EURONEXT 100 index
from 2000-01-04 to 2001-12-31. See examples of
\code{\link{get_rates_from_prices}} for the dataset generation.
}
\keyword{datasets}
