% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply_market_model.R
\name{returns}
\alias{returns}
\title{Constructor for a S3 returns class object}
\usage{
returns(rates, regressor, market_model = c("mean_adj", "mrkt_adj", "sim"),
  estimation_method = c("ols"), estimation_start, estimation_end)
}
\arguments{
\item{rates}{an object of class either \code{zoo} or \code{data.frame}
giving observed rates of returns of secturity.}

\item{regressor}{an object of the same class as \code{rates} representing
rates of returns of the market model, if needed.}

\item{market_model}{a character indicating the market model among
\code{mean_adj}, \code{mrkt_adj}, and \code{sim}.}

\item{estimation_method}{a character, specifying the estimation method for
\code{sim} market model.}

\item{estimation_start}{an object of class Data, giving the start date of
estimation period.}

\item{estimation_end}{an object of class Data, giving the start date of
estimation period.}
}
\value{
An object of S3 class \code{returns}, which contains the following
fields:
\itemize{
\item observed: the object of class \code{zoo}, containing observed rates of
retunrs.
\item predicted: the object of class \code{zoo}, containing predicted by
market model rates of returns.
\item lower95CI: the lower bound of the 95\% Confidence Interval for
predicted rates of returns.
\item upper95CI: the upper bound of the 95\% Confidence Interval for
predicted rates of returns.
\item abnormal: the object of class \code{zoo}, containing abnoraml returns.
\item regressor: the object of class \code{zoo}, containing rates of
regressor (typically market index).
\item market_model: the code name of the market model.
\item full_name_market_model: full name of the market model.
\item estimation_method: the code name of estimation method (applied only for
SIM).
\item full_name_estimation_method: full name of estimation method (applied
only for SIM).
\item coefficients: coeffitients \alpha and \beta for SIM market model
(applied only for SIM).
\item estimation_start: the end of estimation period.
\item estimation_end: the end of estimation period.
\item estimation_length: the length of the estimation period.
}
}
\description{
Constucts an \code{returns} class object with correspoding fields.
}
\details{
The constructor is the generic function, dispatched for classes \code{zoo}
\code{data.frame}. Parameters \code{rates} and \code{regressor} should be
objects of the same class (\code{zoo} or \code{data.frame}). There are three
market model implemented. \code{mean_adj} stands for Mean Adjusted Market
model, which is the avarage of returns during the estimation period.
\code{mrkt_adj} represents Market Adjusted Market model: the securities' rate
of returns are simply market index rates of returns (in terms of parameters -
\code{regressor}). Finally, \code{sim} stands for the Single Index Market
model. For this model only Ordinary Least Squares \code{estimation_method} is
currently implemented. All models are desctibed in Brown and Warner (1985).
}
\references{
Brown S.J., Warner J.B. \emph{Using Daily Stock Returns, The Case
of Event Studies}. Journal of Financial Economics, 14:3-31, 1985.
}

