% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{securities_returns}
\alias{securities_returns}
\title{Returns of seven companies from 2019-04-01 to 2020-04-01}
\format{
A list of eight \code{zoo} elements:
\itemize{
  \item AMZN
  \item ZM
  \item UBER
  \item NFLX
  \item SHOP
  \item FB
  \item UPWK
}
}
\usage{
securities_returns
}
\description{
A list of length seven, elements of which are objects of the class
\code{returns}. The list contains all necessary returns from 2019-04-01 to
2020-04-01 of seven companies, which could profit from COVID-19 lockdown. See
examples of \code{\link{apply_market_model}} for the dataset generation.
}
\keyword{datasets}
