% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{sign_test}
\alias{sign_test}
\title{Simple binomial sign test for event study.}
\usage{
sign_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
The binomial sign test which determines whether the frequency of positive
abnormal returns in the event period is significantly different from one-half.
}
\details{
This test is application of the simple binomial test to the event study,
which indicates whether the cross-sectional frequency of positive abnormal
returns is significantly different from 0.5. This test is stable
to outliers, in other words allows to check if the result is driven by few
companies with extremly large abnomal performance. For this test the
estimation period and event period must not overlap, otherwise an error
occurrs. The test statistic is assumed to have a normal distribution in
approximation under null hypothesis, if the number of securities is large.
Typically is used together with parametric tests. The test is well-specified
for the case, when cross-sectional abnoral returns are not symmetric.
Also this procedure is less sensative to extreme returns than the rank test.
The significance levels of \eqn{\alpha} are 0.1, 0.05, and 0.01
(marked respectively by *, **, and ***).
}
\references{
Boehmer E., Musumeci J., Poulsen A.B. \emph{Event-study
methodology under conditions of event-induced variance}. Journal of Financial
Economics, 30(2):253-272, 1991.
}
\seealso{
\code{\link{nonparametric_tests}}, \code{\link{generalized_sign_test}},
\code{\link{corrado_sign_test}}, \code{\link{rank_test}},
\code{\link{modified_rank_test}}, and \code{\link{wilcoxon_test}}.
}

