% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{t_test}
\alias{t_test}
\title{t-test for event study.}
\usage{
t_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Classical t-test, which examine each date in event window.
}
\details{
Performs the t-test for event study. The procedure of this test is described
in Boehmer 1991, sometimes is called cross-sectional test. Assumes
independence of securities, however is stable to event-induced variance. This
test examines the equality of cross-sectional expected value to zero. The
standard deviation, which is used in this test, is simply cross-section
standard deviation for given day in the event window. It calculates
statistics even if event window and estimation period are overlapped
(intersect). The critical values are Student's t-distributed (no
approximation in limit). The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***).
}
\section{Warning}{
 This test strongly requires cross-sectional independence
and sensitive to the size of the sample.
}

\examples{
## Download the historical prices for ten European insurance companies' stocks
# tickers <- c("ALV.DE", "AML.L", "CS.PA", "ELE.PA", "G.MI", "HNR1.HA",
#              "HSX.L", "MUV2.DE", "RSA.L", "TOP.CO" )
# prices <- get_prices_from_tickers(tickers, start = as.Date("2000-01-01"),
#                                   end = as.Date("2002-01-01"),
#                                   quote = "Close", retclass = "list")
## Estimate the rate of returns form prices
# rates <- get_rates_from_prices(prices, quote = "Close", multi_day = TRUE,
#                                compounding = "continuous")
## Download the prices and estimate the rates of market proxy (index
## ESTX50 EUR P), which is regressor for the sim model
# prices_indx <- get_prices_from_tickers("^STOXX50E",
#                                        start = as.Date("2000-01-01"),
#                                        end = as.Date("2002-01-01"),
#                                        quote = "Close", retclass = "list")
# rates_indx <- get_rates_from_prices(prices_indx, quote = "Close",
#                                     multi_day = TRUE,
#                                     compounding = "continuous")
## Apply Single Index market model
# securities_returns <- apply_market_model(rates = rates,
#                                          regressors = rates_indx,
#                                          same_regressor_for_all = TRUE,
#                                          market_model = "sim",
#                                          estimation_method = "ols",
#                                          estimation_start =
#                                                      as.Date("2001-03-26"),
#                                          estimation_end =
#                                                      as.Date("2001-09-10"))
data(securities_returns)
t_test(securities_returns, as.Date("2001-09-11"), as.Date("2001-09-28"))

}
\references{
Boehmer E., Musumeci J., Poulsen A.B. \emph{Event-study
methodology under conditions of event-induced variance}. Journal of Financial
Economics, 30(2):253-272, 1991.
}
\seealso{
\code{\link{parametric_tests}}, \code{\link{brown_warner_1980}},
\code{\link{brown_warner_1985}}, \code{\link{patell}}, \code{\link{boehmer}},
and \code{\link{lamb}}.
}
