% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parametric_tests.R
\name{t_test}
\alias{t_test}
\title{t-test for event study.}
\usage{
t_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each element
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Classical t-test, which examine each date in event window.
}
\details{
Performs the t-test for event study. The procedure of this test is described
in Boehmer 1991, sometimes is called cross-sectional test. Assumes
independens of securities, however is stable to event-induced variance. This
test examines the equality of cross-sectional expected value to zero. The
standard deviation, which is used in this test, is simply cross-section
standard deviation for given day in the event window. It calculates
statistics even if event window and estimation period are overlapped
(intersect). The critical values are Student's t-destributed (no
approximation in limit). The significance levels of \eqn{\alpha} are 0.1,
0.05, and 0.01 (marked respectively by *, **, and ***).
}
\section{Warning}{
 This test strongly requires cross-sectional independence
and sensative to the size of the sample.
}
\references{
Boehmer E., Musumeci J., Poulsen A.B. \emph{Event-study
methodology under conditions of event-induced variance}. Journal of Financial
Economics, 30(2):253-272, 1991.
}
\seealso{
\code{\link{parametric_tests}}, \code{\link{brown_warner_1980}},
\code{\link{brown_warner_1985}}, \code{\link{patell}}, \code{\link{boehmer}},
and \code{\link{lamb}}.
}

