% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nonparametric_tests.R
\name{wilcoxon_test}
\alias{wilcoxon_test}
\title{Wilcoxon signed rank test for event study.}
\usage{
wilcoxon_test(list_of_returns, event_start, event_end)
}
\arguments{
\item{list_of_returns}{list of objects of S3 class \code{return}, each elemnt
of which is treated as a company.}

\item{event_start}{the object of class \code{Date}, which represents the
first (starting) date of the event window.}

\item{event_end}{the object of class \code{Date}, which represents the last
(ending) date in the event window.}
}
\value{
The table of statistics and significances of the test.
}
\description{
Performs Wilcoxon test on event period for abnormal returns (the latter are
treted as differences).
}
\details{
The estimation periods can overlap with event windows, because the procedure
takes into account only abnormal returns from event windows. The test has the
same algorithm as bulit-in \code{R} \code{\link{wilcox.test}}. The critical
values are exact values, which are obtained from \code{\link{qsignrank}}. The
algorithm is the following: for each day in event window the cross sectional
abnormal returns treated as sample of differences (which actually is true).
Firstly the absolute value of these differences are computed, and
corresponding ranks of non-zero values are calculated. The test statistic is
the sum of ranks, corresponoding to positive abnormal returns. The
significance levels of \eqn{\alpha} are 0.1, 0.05, and 0.01 (marked
respectively by *, **, and ***).
}
\references{
\itemize{
\item Wilcoxon F. \emph{Individual Comparisons by Ranking Mathods}.
Biometrics Bulletin 1(6):80-83, 1945
\item Kolari J.W., Pynnonen S. \emph{Event Study Testing with Cross-sectional
Correlation of Abnormal Returns}. The Review of Financial Studies,
23(11):3996-4025, 2010.
\item Lehmann E.L, \emph{Nonparametrics: Statistical Methods Based on Ranks},
San Francisco: Holden-Day, 1975.
\item Hollander M., Wolfe D.A. \emph{Nonparametric Statistical Methods},
New York: John Wiley & Sons, 1973.
}
}
\seealso{
\code{\link{nonparametric_tests}}, \code{\link{sign_test}},
\code{\link{generalized_sign_test}}, \code{\link{corrado_sign_test}},
\code{\link{rank_test}}, \code{\link{modified_rank_test}}, and
\code{\link{generalized_rank_test}}
}

